+++ modules/itc/src/main/scala/lucuma/itc/ItcWarning.scala
@@ -6,8 +6,4 @@ package lucuma.itc
 import io.circe.Decoder
 import io.circe.generic.semiauto._
 
-final case class ItcWarning(msg: String)
-
-object ItcWarning {
-  implicit val decoder: Decoder[ItcWarning] = deriveDecoder
-}
+final case class ItcWarning(msg: String) derives Decoder
