--- modules/itc/src/main/scala/lucuma/itc/legacy/ItcParameters.scala
@@ -1,17 +1,0 @@
-// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package lucuma.itc.legacy
-
-import io.circe.Encoder
-import io.circe.generic.semiauto._
-import lucuma.itc.ItcObservingConditions
-import lucuma.itc.search.ItcObservationDetails
-
-case class ItcParameters(
-  source:      ItcSourceDefinition,
-  observation: ItcObservationDetails,
-  conditions:  ItcObservingConditions,
-  telescope:   ItcTelescopeDetails,
-  instrument:  ItcInstrumentDetails
-)
