--- modules/itc/src/main/scala/lucuma/itc/legacy/ItcTelescopeDetails.scala
@@ -1,9 +1,0 @@
-// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package lucuma.itc.legacy
-
-import io.circe.Encoder
-import io.circe.Json
-
-case class ItcTelescopeDetails(wfs: ItcWavefrontSensor)
