--- modules/itc/src/main/scala/lucuma/itc/legacy/ItcWavefrontSensor.scala
@@ -1,8 +1,0 @@
-// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package lucuma.itc.legacy
-
-enum ItcWavefrontSensor(val ocs2Tag: String):
-  case PWFS  extends ItcWavefrontSensor("PWFS")
-  case OIWFS extends ItcWavefrontSensor("OIWFS")
